refactor(perf): address CodeRabbit + SonarCloud findings on slice B (#497)
Quality Gate already passed; this clears the open issues:

CodeRabbit
- Added a `file.import` Sentry breadcrumb in cmdAnalyze immediately
  before the import call, matching the project convention for I/O
  operations (CLAUDE.md guidance).
- Text output now prepends `File: <name>` so it mirrors the JSON's
  `file` field — same correlation in CI logs.
- Added explicit #include <QSet> in DrawCallAnalyzer.cpp instead of
  relying on transitive includes.

SonarCloud
- S5276 (size_t → unsigned int): switched the SubEntity iteration in
  analyze() to size_t to match Ogre::Entity::getNumSubEntities()'s
  return type.
- S5276 (qsizetype → int): cast Container::size() at the assignment
  site in mergeSavings() and the uniqueMaterials accumulator.
- S5350: const Ogre::SceneNode* in the analyzeScene scene-graph walk.
- S6004: drawReport now declared inside an if-init-statement in
  MeshInfoOverlay::formatStats.

The remaining MCPServer.cpp:2776 S5817 (`should be const`) is the
same intentional non-const tool method covered by the existing
NOSONAR comment in slice A. MCPServer.cpp:3394 S1116 is a pre-existing
Q_UNUSED(args); in toolGetPivotMode that Sonar mis-attributed to the
new code window — not touched.

+++ src/DrawCallAnalyzer.h
@@ -19,7 +19,9 @@ struct MaterialCluster {
     QStringList entityNames;       // names of entities that use this material
     // Merge potential: every additional entity past the first is a draw call
     // we could save by merging — provided the geometry is compatible.
-    int mergeSavings() const { return entityNames.isEmpty() ? 0 : entityNames.size() - 1; }
+    int mergeSavings() const {
+        return entityNames.isEmpty() ? 0 : static_cast<int>(entityNames.size()) - 1;
+    }
 };
 
 struct MergeSuggestion {
